@@ -29,6 +29,7 @@
     "gatsby-plugin-s3": "^0.3.8",
     "gatsby-plugin-sass": "^5.19.0",
     "gatsby-plugin-sharp": "^4.19.0",
+    "gatsby-plugin-feed": "^4.22.0",
     "gatsby-plugin-sitemap": "^5.19.0",
     "gatsby-plugin-styled-components": "^5.19.0",
     "gatsby-plugin-typescript": "^4.19.0",
@@ -62,7 +63,7 @@
     "schema-dts": "^0.8.2",
     "styled-components": "^5.3.5",
     "swiper": "^8.3.2",
-    "typescript": "^4.3.5"
+    "typescript": "^4.8.3"
   },
   "devDependencies": {
     "@types/classnames": "^2.2.10",
