@@ -1,6 +1,6 @@
-import { siteMetadata } from '../../gatsby-config'
+import config from '../../gatsby-config'
 import React from 'react'
 
 export function descriptionOrDefault(description: string | undefined){
-  return description ?? siteMetadata.description
+  return description ?? config.siteMetadata.description
 }
