@@ -1,8 +1,8 @@
-import { siteMetadata } from '../../gatsby-config'
+import config from '../../gatsby-config'
 
 export function resolveUrl(pathname: string) {
   if (pathname.startsWith('http://') || pathname.startsWith('https://')) {
     return pathname
   }
-  return new URL(pathname, siteMetadata.siteUrl).toString()
+  return new URL(pathname, config.siteMetadata.siteUrl).toString()
 }
